requirements.txt
customtkinter>=5.2.0
Pillow>=10.0.0
spotipy>=2.23.0
yt-dlp>=2024.04.09
